# api/openapi.3.1.projection.yaml
...
openapi: 3.1.0
info:
  title: Bedrock API
  version: 2.0.0
  description: |
    Bedrock 2.0 HTTP API. Envelope `{ code, message, data? }`.
    Pagination data shape: `{ items, total, page, page_size, total_pages }`.
servers:
  - url: /api/v1
    description: Relative to the Bedrock host
security:
  - bearerAuth: []
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users
      operationId: usersList
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserPage"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
    post:
      tags: [Users]
      summary: Create user
      operationId: usersCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /users/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [Users]
      summary: Get user
      operationId: usersGet
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
    put:
      tags: [Users]
      summary: Update user
      operationId: usersUpdate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Users]
      summary: Delete user
      operationId: usersDelete
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
...
  /dictionaries:
    get:
      tags: [Dictionaries]
      summary: List dictionaries
      operationId: dictionariesList
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/DictionaryPage"
        "403":
          $ref: "#/components/responses/Error"
    post:
      tags: [Dictionaries]
      summary: Create dictionary
      operationId: dictionariesCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DictionaryCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Dictionary"
        "400":
          $ref: "#/components/responses/Error"
  /dictionaries/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [Dictionaries]
      summary: Get dictionary
      operationId: dictionariesGet
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Dictionary"
        "404":
          $ref: "#/components/responses/Error"
    put:
      tags: [Dictionaries]
      summary: Update dictionary
      operationId: dictionariesUpdate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DictionaryUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Dictionary"
        "400":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Dictionaries]
      summary: Delete dictionary
      operationId: dictionariesDelete
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /operation-logs:
    get:
      tags: [OperationLogs]
      summary: Query operation logs
      operationId: operationLogsList
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: user_id
          in: query
          schema:
            type: integer
        - name: action
          in: query
          schema:
            type: string
        - name: resource_type
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/OperationLogPage"
        "403":
          $ref: "#/components/responses/Error"
...
  /build-runs/{id}/log:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [BuildRuns]
      summary: Fetch build log text (WS preferred for live streaming)
      operationId: buildRunsLog
      responses:
        "200":
          description: Plain text log
          content:
            text/plain:
              schema:
                type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ID:
      name: id
      in: path
      required: true
      schema:
        type: integer
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    SuccessEnvelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string
        request_id:
          type: string
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        display_name:
          type: string
        email:
          type: string
        avatar:
          type: string
        is_active:
          type: boolean
        is_super_admin:
          type: boolean
        role_ids:
          type: array
          items:
            type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserCreateRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        display_name:
          type: string
        email:
          type: string
        is_active:
          type: boolean
        role_ids:
          type: array
          items:
            type: integer
    UserUpdateRequest:
      type: object
      properties:
        display_name:
          type: string
        email:
          type: string
        password:
          type: string
        is_active:
          type: boolean
        role_ids:
          type: array
          items:
            type: integer
...
    Dictionary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        description:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/DictItem"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DictItem:
      type: object
      properties:
        id:
          type: integer
        dictionary_id:
          type: integer
        label:
          type: string
        value:
          type: string
        sort_order:
          type: integer
        enabled:
          type: boolean
    DictionaryCreateRequest:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/DictItem"
    DictionaryUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/DictItem"
    OperationLog:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        username:
          type: string
        action:
          type: string
        resource_type:
          type: string
        resource_id:
          type: string
        details:
          type: string
        ip_address:
          type: string
        created_at:
          type: string
          format: date-time
    Page:
      type: object
      required: [items, total, page, page_size, total_pages]
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
          format: int64
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer
    UserPage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/User"
...
    DictionaryPage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Dictionary"
    OperationLogPage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OperationLog"
...